mutex and multiple tx moved to module for a channel
